preinstall: move errors to their own file
