+++ efi/preinstall/errors.go
@@ -1,0 +1,77 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+
+/*
+ * Copyright (C) 2024 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package preinstall
+
+import "errors"
+
+// NoHardwareRootOfTrustError is returned wrapped from [RunChecks] if the platform
+// firmware is not protected by a hardware root-of-trust. This won't be returned if
+// the PermitVirtualMachine flag is supplied to [RunChecks] and the current
+// environment is a virtual machine.
+type NoHardwareRootOfTrustError struct {
+	err error
+}
+
+func (e *NoHardwareRootOfTrustError) Error() string {
+	return "no hardware root-of-trust properly configured: " + e.err.Error()
+}
+
+func (e *NoHardwareRootOfTrustError) Unwrap() error {
+	return e.err
+}
+
+// UnsupportedPlatformError is returned wrapped from [RunChecks] if this platform
+// is not supported for FDE.
+type UnsupportedPlatformError struct {
+	err error
+}
+
+func (e *UnsupportedPlatformError) Error() string {
+	return "unsupported platform: " + e.err.Error()
+}
+
+func (e *UnsupportedPlatformError) Unwrap() error {
+	return e.err
+}
+
+var (
+	// ErrCPUDebuggingNotLocked is returned wrapped from RunChecks if the CPU
+	// has silicon debugging features but these have not been disabled and
+	// locked by the platform firmware. This won't be returned if the
+	// PermitVirtualMachine flag is supplied to RunChecks and the current
+	// environment is a virtual machine.
+	ErrCPUDebuggingNotLocked = errors.New("CPU debugging features are not disabled and locked")
+
+	// ErrInsufficientDMAProtection is returned wrapped from RunChecks if the platform
+	// firmware indicates that I/O DMA protection was disabled at some point. This won't
+	// be returned if the PermitVirtualMachine flag is supplied to RunChecks and the current
+	// environment is a virtual machine.
+	ErrInsufficientDMAProtection = errors.New("the platform firmware indicates that DMA protections are insufficient")
+
+	// ErrNoKernelIOMMU is returned wrapped from RunChecks if there is no IOMMU active. This
+	// won't be returned if the PermitVirtualMachine flag is supplied to RunChecks and
+	// the current environment is a virtual machine.
+	ErrNoKernelIOMMU = errors.New("no kernel IOMMU support was detected")
+
+	// ErrUEFIDebuggingEnabled is returned wrapped from RunChecks if the platform firmware
+	// has a debugging endpoint enabled. This won't be returned if the PermitVirtualMachine
+	// flag is supplied to RunChecks and the current environment is a virtual machine.
+	ErrUEFIDebuggingEnabled = errors.New("the platform firmware contains a debugging endpoint enabled")
+)
